Struct bencode::ByteStr

pub struct ByteStr {
    // some fields omitted
}

A borrowed byte string.

This wrapper is necessary for a byte string to be encoded as a string in the bencode format rather than as a list.

Methods

impl ByteStr

fn from_bytes(b: &[u8]) -> &ByteStr

Returns a byte slice as a &ByteStr.

fn as_bytes(&self) -> &[u8]

Returns the byte string as a slice.

fn as_str(&self) -> Option<&str>

Returns the byte string as a &str, if it contains only valid UTF-8.