Struct bencode::Decoder

pub struct Decoder<'a> {
    // some fields omitted
}

Decodes values from a stream of bytes.

Methods

impl<'a> Decoder<'a>

fn new(data: &[u8]) -> Decoder

Constructs a new Decoder, reading from the given byte string.

fn remaining(&self) -> usize

Returns the number of bytes remaining in the stream.

fn position(&self) -> u64

Returns the current position of the cursor.

fn set_position(&mut self, pos: u64)

Sets the current position of the cursor.

fn finish(self) -> Result<(), DecodeError>

Returns an error if there is data remaining in the stream.

fn read(&mut self, buf: &mut [u8]) -> Result<(), DecodeError>

Reads a series of bytes from the stream equal to buf.len(). If fewer bytes are available to read, an error is returned.

fn read_byte(&mut self) -> Result<u8, DecodeError>

Reads a single byte from the stream. If no bytes are available to read, an error is returned.

fn peek_byte(&self) -> Result<u8, DecodeError>

Reads a single byte from the stream without advancing the cursor.

fn peek_bytes(&self, n: usize) -> Result<&[u8], DecodeError>

Returns a slice of bytes without advancing the cursor. If fewer than n bytes are available, an error is returned.

fn read_integer<T: Integer>(&mut self) -> Result<T, DecodeError>

Reads an integer value from the stream.

fn read_number<T: Integer>(&mut self) -> Result<T, DecodeError>

Reads a number from the stream. This does not include the i prefix and e suffix.

fn read_bytes(&mut self) -> Result<Vec<u8>, DecodeError>

Reads a byte string from the stream.

fn read_str(&mut self) -> Result<String, DecodeError>

Reads a UTF-8 encoded string from the stream.

fn read_dict<T: Decodable>(&mut self) -> Result<BTreeMap<String, T>, DecodeError>

Reads a key value mapping from the stream.

fn read_list<T: Decodable>(&mut self) -> Result<Vec<T>, DecodeError>

Reads a series of values from the stream.

fn read_struct<T, F>(&mut self, f: F) -> Result<T, DecodeError> where F: FnOnce(&mut Decoder<'a>) -> Result<T, DecodeError>

Reads a key value mapping from the stream as a struct.

The given callable is expected to call read_field for each field and read_option for any optional fields, in lexicographical order.

fn read_field<T: Decodable>(&mut self, name: &str) -> Result<T, DecodeError>

Reads a single field from the stream.

fn read_option<T: Decodable>(&mut self, name: &str) -> Result<Option<T>, DecodeError>

Reads an optional field from the stream.

fn skip_item(&mut self) -> Result<(), DecodeError>

Advances the cursor beyond the current value.

fn skip(&mut self, n: usize) -> Result<(), DecodeError>

Advances the cursor n bytes.

fn skip_while<P>(&mut self, predicate: P) -> Result<(), DecodeError> where P: FnMut(u8) -> bool

Advance bytes in the stream until predicate returns false.

fn read_while<P>(&mut self, predicate: P) -> Result<Vec<u8>, DecodeError> where P: FnMut(u8) -> bool

Reads bytes from the stream until predicate returns false.

fn expect(&mut self, byte: u8) -> Result<(), DecodeError>

Returns an error if the next byte is not byte.