Struct bencode::Encoder

pub struct Encoder {
    // some fields omitted
}

Encodes values into a stream of bytes.

Methods

impl Encoder

fn new() -> Encoder

Constructs a new Encoder.

fn into_bytes(self) -> Vec<u8>

Consumes the Encoder and returns the encoded bytes.

fn write_byte(&mut self, b: u8) -> Result<(), EncodeError>

Writes a single byte to the stream.

fn write(&mut self, b: &[u8]) -> Result<(), EncodeError>

Writes a series of bytes to the stream.

fn write_integer<T: Integer>(&mut self, t: T) -> Result<(), EncodeError>

Writes an integer value to the stream.

fn write_number<T: Integer>(&mut self, t: T) -> Result<(), EncodeError>

Writes a number to the stream. This does not include i prefix and e suffix.

fn write_bytes(&mut self, b: &[u8]) -> Result<(), EncodeError>

Writes a byte string to the stream.

fn write_str(&mut self, s: &str) -> Result<(), EncodeError>

Writes a UTF-8 encoded string to the stream.

fn write_dict<K, V>(&mut self, map: &BTreeMap<K, V>) -> Result<(), EncodeError> where K: Ord + AsRef<str>, V: Encodable

Writes a key value mapping to the stream.

fn write_list<T: Encodable>(&mut self, t: &[T]) -> Result<(), EncodeError>

fn write_struct<F>(&mut self, f: F) -> Result<(), EncodeError> where F: FnOnce(&mut Encoder) -> Result<(), EncodeError>

Writes a key value mapping from a struct to the stream.

The given callable is expected to call write_field for each field and write_option for any optional fields, in lexicographical order.

fn write_field<T: ?Sized + Encodable>(&mut self, name: &str, t: &T) -> Result<(), EncodeError>

Writes a single field to the stream.

fn write_option<T: Encodable>(&mut self, name: &str, t: &Option<T>) -> Result<(), EncodeError>

Writes an optional field to the stream.